refactor(database): move supabase information and connection to 'database' directory
